Hello world added.
Hello world test program to make sure everyone is on the same page added
as described in the class. Currently the program has no build
architecture and must be built manually from source using something like
gcc.

This commit also adds an important file, .gitignore, which we should
keep up to date to keep generated garbage out of source control since
it's not source or documentation.