Fixed broken build caused by set CXXFLAGS
If CXXFLAGS is set in the environment by the user then the build would
break.  Now configure checks if CXXFLAGS is already set by assuming that
if the env has a string object named CXXFLAGS then the user set it.  If
that is the case it is converted to a list.