Clarified project goals in proposal document.
Removed "Learning Goals":  we have one goal, and that is to deliver one
heck of a YAML library.  I lowered my expectations given time
constraints, so that our goals are to implement a specific subset of the
1.2 standard:  scalars, sequences and mappings.

Also, changed the last prioirity from "feature-rich" to "quality code":
otherwise it would conflict with the first requirement, of delivering a
subset of the YAML specs.  "Quality code" means to code well so that it
is feasible to fill in the missing specs later on if desired.