string scalars can span across multiple lines
This just in:  string scalars can span across multiple lines!  Newlines
in string scalars are never significant, though, and any newline is
always folded into a single space character.  This as per section 2.3 of
the standard.

You may ask what is the different between scalars and literal style or
folded style, then.  For what I understand, literal style's newlines are
always significant, and folded style's newlines are significant only
when ending empty or whitespace-only lines.  This as opposed to string
scalars' newlines, which are never significant and thus are always
folded into (replaced by) a single space.  Newlines that are significant
aren't folded.