Added a demo application
We'll use this for demonstrating our library.  We might as well use it
for development purposes, as a playground.  waf is configured to build
our project as a library, not an executable, so by having a demo area we
can actually build an executable we can build and run and test.