Added build instructions to README
Also added directory structure description.  I state that demo is in
demo/, not src/demo/, but I plan on moving demo from src/demo/ to the
root directory.

I mention that it is possible to build the library manually, but haven't
provided the details on how to do that.  I only specified the details on
building with Boost.Build.  Still need to add the details on how to
build manually ..