each branch in scan() returns
as opposed to set the return value which will ultimately be returned.

the reason is that in the previous approach i had to use if-else-if so
that when a branch is entered no other branch will be entered.  but as
it turns out sometimes i can enter a branch and then change my mind
and let other branches have a chance to catch the condition.

for example, if i scan a '-' it may be a start of a negative number, or
it may be a document-begin, etc.  i need to get into a branch, and then
decide if i should give another branch a chance (by restoring the input
first of course).