created presentation_input, to hold yaml stream
this to extract the yaml stream from the scanner.  then, i plan on
passing a reference to the stream to individual token scanner objects.
each token scanner will then have access to the input stream, so it can
look at peek, put extracted characters back into the buffer and so on.