collapsed scan(), recognize() into one scan()
Replaced bool scan() and virtual bool recognize() from token_scanner
into virtual token scan() that may throw.  The old mechanism just plain
doesn't work: when a scanner would recognize a token it would store it
for ITSELF using the previous() function.  The head of the list which
ignited the call to scan() would never know what infact got lexed,
because the lexed token remained in the scanner object the recognized
the token, never propagated back the list.

So now that the virtual function returns the lexed token, the lexed
token is properly propagated back to the original token_scanner that
ignited the lexing.

The non-virtual scan() function is not needed because finally it really
doesn't do anything:  in the old scheme, a scanner would say it couldn't
recognize the token by returning false.  But in this case, it might
instead just call token_scanner::scan(), which just forwards the request
to the successor.  Exactly the same thing is being done -- but in fact
with fewer lines of code.

Because the success of the lexing is not specified anymore by a boolean
return value, I had to find a way to signal failure in lexing.  At the
moment this is done by throwing a std::logic_error, bur I recognize this
is not a perfect solution at all, because there is nothing excpetional
about not recognizing a bunch of characters as valid tokens.  This is to
be fixed.

In this commit I also introduce a new practice:  making friends with
test classes.  I can't think of a better way to allow the tests to test
a class down to the bones (where private parts are), but I'm open for
ideas.  In the meantime, token_scanner_test is token_scanner's friend.

Also in this commit, a very long commit message.