fix:  lex sequence elements
a sequence element separator is a comma followed by a space.  the
following:

 a, b

is a string and should be lexed as such.  however, the following:

 [a, b]

is a sequence, and should be lexed as such.  therefore, we have to keep
track of sequence_begin tokens we see, to know if a ", " denotes an
element separator or just part of a string.