lex for token::DOCUMENT_BEGIN
Must be "---" followed by whitespace.  The following have DOCUMENT_BEGIN
in them:

---
hello, world

--- hello, world

However the following have no DOCUMENT_BEGIN in them but are plain
scalar strings that include "---" in the string:

---hello, world

---

(a document that has nothing other than "---", not even a single space
or a single newline after the last "-" character).