parser:  allow integers when strings are allowed
this is mostly a hack, as I assume token::INTEGER holds a string as its
value, and I use string_node to represent the node.  thus, from the
representation perspective, there is no difference between a string and
an integer.  it might actually be a Good Thing in a broad sense, but the
scalar node's name should probably be different from string_node, and I
should be able to extract an integer out of an integer scalar (for that
matter, or an object T out of a scalar, if and when I support tags).