added event-dispatching mechanism to parser
when a production is parsed an event is fired.  the following
productions fire an event:

 * document
 * {sequence,mapping}_{begin,end}
 * pair
 * string
 * integer
 * eos

the parser's ctor is modified for this purpose;  it now takes an
event_handler.  to catch the events:

 1. subclass from event_handler
 2. override the handlers of the events you wish to handle
 3. pass your handler object to the parser ctor

the parser assumes ownership of the handler, so be careful not to delete
the handler you instantiate.

example:

class my_handler : public event_handler {
public:
    void on_integer(const string& number) { /* do stuff. */ }
    // on_string(const string& str), on_document(), on_eos() etc.
};

void f(istream& is)
{
    my_handler* h = new my_handler();
    parser p(is, h);

    // p assumes ownership of h, so don't delete h.
}