demo:  be more unixy, not interactive
it is easy to wrap the binary with a shell loop if desired.  but if the
application is interactive then it isn't very useable in a pipe.
specifically, in the state it was prior to this patch cin would clear()
after each iteration, thus rendering an infinite loop when stdin doesn't
come from a tty.  also, the application would print out a useless
greeting message when stdin isn't connected to a tty.