allow input from stdin in configuration demo
if configuration demo is invoked with no command-line argument or with
argument "-", input is taken from stdin.  Otherwise the first
command-line argument is taken as the name of the file to open for
parsing.

This implies I had to take off the usage message.  Instead, I placed it
in the source file.  I think this is fine because the whole point of
this is demonstrating libyaml++, so the source file will be read anyhow.