handle "configuration file syntax errros"
this obviously might not work, but it's just for the sake of
demonstration.  the check relies on mapping every string or integer
event to the next.  it works for mappings, but in the case of a sequence
it will still map every two elements of the sequence as key-value pairs,
and report an error if there is an odd number of elements.  similarly,
if the document consists of a single scalar an error will be emitted
(althought that might actually be more reasonable).